Serve arbitrary urls from the current directory as well
Previously, only github-style absolute urls (containing /raw/) were
served from the current directory and all others were served from the
module directory.

With this change, only the handful of urls that actually exist in the
module directory are served from there, all others are served from the
current directory. This should make sure that linking to images or
other resources in a markdown document works as you expect it to.

Any /raw/ urls are still detected and modified to still work (stripping
the fixed prefix, leaving only the relative link) as well.

This commit is based on #11 by David Bankier <dbankier@hotmail.com>, but
pretty much changes all of the code.